refactor: rename tooltip as geometry-tooltip
