+++ src/interaction/action/component/geometry-tooltip.ts
